+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2]
+### Added
+- Supported command `doctrine:migrations:current`
+- Supported command `doctrine:migrations:latest`
+- Supported command `doctrine:migrations:list`
+- Supported command `doctrine:migrations:status`
+- Supported command `doctrine:migrations:sync-metadata-storage`
+- Supported command `doctrine:migrations:up-to-date`
+
+### Changed
+- Internal class renaming
+- Command description as the original one in `doctrine:migrations:diff` and `doctrine:migrations:version`
+
 ## [0.1.1]
 ### Added
 - Project cleanup